Add export() to inductiva.storage
